moved config helper to util package
